+++ trunk/src/headers/gx_main_interface.h
@@ -161,10 +161,10 @@ namespace gx_gui
     virtual void addregler(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addbigregler(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addslider(const char* label, float* zone, float init, float min, float max, float step);
-    virtual void addtoggle(const char* label, float* zone);
+    virtual void addtoggle(const char* label, int* zone);
     virtual void addbtoggle(const char* label, float* zone);
-    virtual void addswitch(const char* label, float* zone);
-    virtual void addminiswitch(const char* label, float* zone);
+    virtual void addswitch(const char* label, int* zone);
+    virtual void addminiswitch(const char* label, int* zone);
     virtual void addNumEntry(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addNumDisplay(const char* label, float* zone);
     virtual void addLiveWaveDisplay(const char* label, float* zone , float* zone1);
